# SmartPowerElectricAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SmartPowerElectricAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Factura/create/{idOrden}": {
      "post": {
        "tags": [
          "Factura"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idOrden",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FacturaDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Factura/{id}": {
      "delete": {
        "tags": [
          "Factura"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Factura"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FacturaDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Factura/sendInvoice/{idFactura}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idFactura",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Factura/downloadInvoice/{idFactura}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idFactura",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Material/create/{idOrden}": {
      "post": {
        "tags": [
          "Material"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idOrden",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MaterialDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Material/{id}": {
      "delete": {
        "tags": [
          "Material"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Material"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/MaterialDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "FacturaDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "montoACobrar": {
          "format": "double",
          "type": "number"
        },
        "numeroFactura": {
          "type": "string"
        },
        "emailEnviado": {
          "type": "boolean"
        },
        "facturaCompletada": {
          "type": "boolean"
        },
        "idOrden": {
          "format": "int32",
          "type": "integer"
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "MaterialDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "cantidad": {
          "format": "double",
          "type": "number"
        },
        "precioTotal": {
          "format": "double",
          "type": "number"
        },
        "idTipoMaterial": {
          "format": "int32",
          "type": "integer"
        },
        "tipoMaterial": {
          "type": "string"
        },
        "idUnidadMedida": {
          "format": "int32",
          "type": "integer"
        },
        "unidadMedida": {
          "type": "string"
        },
        "idOrden": {
          "format": "int32",
          "type": "integer"
        },
        "fechaCreacion": {
          "type": "string"
        },
        "fechaEliminado": {
          "type": "string"
        },
        "eliminado": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
  }
}